requirements.txt
paddleocr==2.7.2
paddlepaddle==2.6.2
opencv-python==4.6.0.66
pillow==10.3.0
mss==9.0.1
pywin32==306
keyboard==0.13.5
pystray==0.19.5
pygetwindow==0.0.9
llama-cpp-python==0.2.72
pyperclip==1.8.2
